@@ -1,9 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+	
     <TargetFramework>netstandard2.1</TargetFramework>
     <PackageId>NeoCortexApi</PackageId>
-    <Version>1.1.3</Version>
+    <Version>1.1.4</Version>
     <Authors>Damir Dobric</Authors>
     <Company>daenet GmbH</Company>
     <Product>Hierarchical Temporal Framework in C#</Product>
@@ -33,7 +34,7 @@
   </ItemGroup>
   
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <LangVersion>7.3</LangVersion>
+	  <LangVersion>8.0</LangVersion>
     <DefineConstants>TRACE;USE_AKKA</DefineConstants>
     <DebugType>portable</DebugType>
     <DebugSymbols>true</DebugSymbols>
@@ -45,7 +46,7 @@
   </PropertyGroup>
 
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <LangVersion>7.3</LangVersion>
+	  <LangVersion>8.0</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="Anomaly.cs" />
@@ -53,6 +54,10 @@
     <Compile Remove="DistributedCompute\HtmSparseIntDictionary.cs" />
     <Compile Remove="Encoders\EncoderBuilder.cs" />
     <Compile Remove="HtmSerializer2.cs" />
+    <Compile Remove="NAA\ComputeCycleInput.cs" />
+    <Compile Remove="NAA\ComputeState.cs" />
+    <Compile Remove="NAA\CorticalArea.cs" />
+    <Compile Remove="NAA\NeuralAssociationsAlgorithm.cs" />
     <Compile Remove="Network\NearestClassifier.cs" />
   </ItemGroup>
 
@@ -67,6 +72,10 @@
     <None Include="DistributedCompute\HtmSparseIntDictionary.cs" />
     <None Include="Encoders\EncoderBuilder.cs" />
     <None Include="HtmSerializer2.cs" />
+    <None Include="NAA\ComputeCycleInput.cs" />
+    <None Include="NAA\ComputeState.cs" />
+    <None Include="NAA\CorticalArea.cs" />
+    <None Include="NAA\NeuralAssociationsAlgorithm.cs" />
     <None Include="NeoCortexApi\NeoCortexAPI.png">
       <Pack>True</Pack>
       <PackagePath></PackagePath>
