@@ -161,12 +161,21 @@ namespace NeoCortexApi.Classifiers
         /// </summary>
         /// <param name="input">Any kind of input.</param>
         /// <param name="output">The SDR of the input as calculated by SP.</param>
-        /// <param name="predictedOutput"></param>
         public void Learn(TIN input, Cell[] output)
         {
             Stopwatch sw = Stopwatch.StartNew();
             var cellIndicies = GetCellIndicies(output);
 
+            Learn(input, cellIndicies);
+        }
+
+        /// <summary>
+        /// Assotiate specified input to the given set of predictive cells. This can also be used to classify Spatial Pooler Columns output as int array
+        /// </summary>
+        /// <param name="input">Any kind of input.</param>
+        /// <param name="output">The SDR of the input as calculated by SP as int array</param>
+        public void Learn(TIN input, int[] cellIndicies)
+        {
             if (m_AllInputs.ContainsKey(input) == false)
                 m_AllInputs.Add(input, new List<int[]>());
 
@@ -319,21 +328,32 @@ namespace NeoCortexApi.Classifiers
         /// <returns>List of predicted values with their similarities.</returns>
         public List<ClassifierResult<TIN>> GetPredictedInputValues(Cell[] predictiveCells, short howMany = 1)
         {
+            var cellIndicies = GetCellIndicies(predictiveCells);
+
+            return GetPredictedInputValues(cellIndicies, howMany);
+        }
+
+        /// <summary>
+        /// Gets multiple predicted values. This can also be used to classify Spatial Pooler Columns output as int array
+        /// </summary>
+        /// <param name="predictiveCells">The current set of predictive cells in int array.</param>
+        /// <param name="howMany">The number of predections to return.</param>
+        /// <returns>List of predicted values with their similarities.</returns>
+        public List<ClassifierResult<TIN>> GetPredictedInputValues(int[] cellIndicies, short howMany = 1)
+        {
             List<ClassifierResult<TIN>> res = new List<ClassifierResult<TIN>>();
             double maxSameBits = 0;
             TIN predictedValue = default;
             Dictionary<TIN, ClassifierResult<TIN>> dict = new Dictionary<TIN, ClassifierResult<TIN>>();
 
             var predictedList = new List<KeyValuePair<double, string>>();
-            if (predictiveCells.Length != 0)
+            if (cellIndicies.Length != 0)
             {
                 int indxOfMatchingInp = 0;
-                Debug.WriteLine($"Item length: {predictiveCells.Length}\t Items: {this.m_AllInputs.Keys.Count}");
+                Debug.WriteLine($"Item length: {cellIndicies.Length}\t Items: {this.m_AllInputs.Keys.Count}");
                 int n = 0;
 
                 List<int> sortedMatches = new List<int>();
-
-                var cellIndicies = GetCellIndicies(predictiveCells);
 
                 Debug.WriteLine($"Predictive cells: {cellIndicies.Length} \t {Helpers.StringifyVector(cellIndicies)}");
 
@@ -700,13 +720,14 @@ namespace NeoCortexApi.Classifiers
         /// <summary>
         /// Traces out all cell indicies grouped by input value.
         /// </summary>
-        public void TraceState(string fileName = null)
+        public string TraceState(string fileName = null)
         {
+            StringWriter strSw = new StringWriter();
+
             StreamWriter sw = null;
+
             if (fileName != null)
                 sw = new StreamWriter(fileName);
-            else
-                sw = new StreamWriter(fileName.Replace(".csv", "HtmClassifier.state.csv"));
 
             List<TIN> processedValues = new List<TIN>();
 
@@ -714,36 +735,37 @@ namespace NeoCortexApi.Classifiers
             // Trace out the last stored state.
             foreach (var item in this.m_AllInputs)
             {
-                Debug.WriteLine("");
-                Debug.WriteLine($"{item.Key}");
-                Debug.WriteLine($"{Helpers.StringifyVector(item.Value.Last())}");
+                strSw.WriteLine("");
+                strSw.WriteLine($"{item.Key}");
+                strSw.WriteLine($"{Helpers.StringifyVector(item.Value.Last())}");
+            }
+
+            strSw.WriteLine("........... Cell State .............");
+
+            foreach (var item in m_AllInputs)
+            {
+                strSw.WriteLine("");
+
+                strSw.WriteLine($"{item.Key}");
+
+                strSw.Write(Helpers.StringifySdr(new List<int[]>(item.Value)));
+
+                //foreach (var cellState in item.Value)
+                //{
+                //    var str = Helpers.StringifySdr(cellState);
+                //    strSw.WriteLine(str);
+                //}
             }
 
             if (sw != null)
             {
+                sw.Write(strSw.ToString());
                 sw.Flush();
                 sw.Close();
             }
 
-            Debug.WriteLine("........... Cell State .............");
-
-            using (var cellStateSw = new StreamWriter(fileName.Replace(".csv", "HtmClassifier.fullstate.csv")))
-            {
-                foreach (var item in m_AllInputs)
-                {
-                    Debug.WriteLine("");
-                    Debug.WriteLine($"{item.Key}");
-
-                    cellStateSw.WriteLine("");
-                    cellStateSw.WriteLine($"{item.Key}");
-                    foreach (var cellState in item.Value)
-                    {
-                        var str = Helpers.StringifyVector(cellState);
-                        Debug.WriteLine(str);
-                        cellStateSw.WriteLine(str);
-                    }
-                }
-            }
+            Debug.WriteLine(strSw.ToString());
+            return strSw.ToString();
         }
 
 
