@@ -32,25 +32,132 @@ namespace NeoCortexApi
             return vector;
         }
 
-        public string TraceSdrsMethod(List<int[]> indexesOfNoneZeros, string paddingTxt = " ", bool renderSimilarityMatrix = false)
-        { 
-            throw new NotImplementedException();
+        #region TraceSDR
+
+        
+        /// <summary>
+        /// Creates string representation from one dimensional value. 
+        /// </summary>
+        /// <see cref=""/>
+        /// <param name="sdrs">the SDR sets</param>
+        /// <returns>string of traced output SDRs</returns>
+        public static string StringifySdr(List<int[]> sdrs)
+        {
+            //List of string of arrays for SDR set
+            var heads = new List<int>(new int[sdrs.Count]);
+
+            //The count for SDR starting from initial position [0,0]
+            var outputs = new StringBuilder[sdrs.Count];
+
+            while (true)
+            {
+                //We set the minimum value as initial value of SDRs can be 0
+                int minActiveColumn = -1;
+
+                minActiveColumn = ArrangeSdr(sdrs, heads, minActiveColumn);
+
+                if (minActiveColumn == -1)
+                {
+                    //Stores a mutable string of characters.
+                    var result = new StringBuilder();
+
+                    foreach (var output in outputs)
+                    {
+                        result.AppendLine(output.ToString());
+                    }
+                    return result.ToString();
+                }
+
+                Append_ActiveColumn(sdrs, heads, outputs, minActiveColumn);
+            }
         }
 
+
+        /// <summary>
+        /// Stores the SDR values from both sets and arrange them.
+        /// </summary>
+        /// <param name="sdrs">The SDR values as index bit of active neurons</param>
+        /// <param name="heads">The two representations taken as input.</param>
+        /// <param name="outputs">Represents every index bit of representations.</param>
+        /// <param name="minActiveColumn">Stores the similar semantic active and inactive bits for comparison.</param>
+        /// <summary>
+        public static void Append_ActiveColumn(List<int[]> sdrs, List<int> heads, StringBuilder[] outputs, int minActiveColumn)
+        {
+            for (int i = 0; i < sdrs.Count; i++)
+            {
+                if (outputs[i] == null)
+                {
+                    outputs[i] = new StringBuilder();
+                }
+                var head = heads[i];
+                var sdr = sdrs[i];
+
+                if (head < sdr.Length && sdr[head] == minActiveColumn)
+                {
+                    outputs[i].Append(minActiveColumn);
+                    outputs[i].Append(", ");
+                    heads[i] = head + 1;
+                }
+                else
+                {
+                    //Creates a padding or spacing on the indices of inactive bits
+                    var numOfSpaces = minActiveColumn.ToString().Length;
+                    for (var j = 0; j < numOfSpaces; j++)
+                    {
+                        outputs[i].Append(" ");
+                    }
+                    outputs[i].Append(", ");
+                }
+            }
+        }
+
+        /// <summary>
+        /// Creates the results of SDRs in a well arrangement
+        /// </summary>
+        private static int ArrangeSdr(List<int[]> sdrs, List<int> heads, int minActiveColumn)
+        {
+            for (int i = 0; i < sdrs.Count; i++)
+            {
+                var head = heads[i];
+                var sdr = sdrs[i];
+
+                if (heads[i] > sdr.Length - 1)
+                {
+                    continue;
+                }
+
+                var activeColumn = sdr[head];
+                if (minActiveColumn == -1)
+                {
+                    minActiveColumn = activeColumn;
+                }
+                else
+                {
+                    if (activeColumn < minActiveColumn)
+                    {
+                        minActiveColumn = activeColumn;
+                    }
+                }
+            }
+
+            return minActiveColumn;
+        }
+
+        #endregion
         /// <summary>
         /// Creates string representation from one dimensional vector.
         /// </summary>
         /// <param name="vector"></param>
         /// <param name="separator">The separator used between bits.</param>
         /// <returns></returns>
-        public static string StringifyVector(int[] vector, string separator=", ")
+        public static string StringifyVector(int[] vector, string separator = ", ")
         {
             StringBuilder sb = new StringBuilder();
 
             foreach (var vectorBit in vector)
             {
                 sb.Append(vectorBit);
-                if(separator != null)
+                if (separator != null)
                     sb.Append(separator);
             }
 
@@ -250,10 +357,10 @@ namespace NeoCortexApi
         /// <param name="matrix">Two-dimensianal matrix of cross-similarities between given values. Values are already rendered as string.</param>
         /// <returns></returns>
 
-        public static string RenderSimilarityMatrix( string[] values, string[,] matrix)
+        public static string RenderSimilarityMatrix(string[] values, string[,] matrix)
         {
             System.IO.StringWriter sw = new System.IO.StringWriter();
-            
+
             sw.Write($"{string.Format(" {0,-15}", "")} |");
 
             for (int k = 0; k < values.Length; k++)
