@@ -18,7 +18,12 @@ namespace NeoCortexApi
     /// and switches-off the boosting mechanism (new-born effect) after the SP has entered a stable state 
     /// for all seen input patterns.
     /// </summary>
-    public class HomeostaticPlasticityController
+    /// <remarks>
+    /// The research related to this component can be found here:
+    /// https://www.researchgate.net/publication/358996456_On_the_Importance_of_the_Newborn_Stage_When_Learning_Patterns_with_the_Spatial_Pooler
+    /// Published 2022 in Springer Nature Computer Sciences.
+    /// </remarks>
+    public class HomeostaticPlasticityController : ISerializable
     {
         private double m_RequiredSimilarityThreshold;
 
@@ -59,6 +64,7 @@ namespace NeoCortexApi
         /// </summary>
         private Action<bool, int, double, int> m_OnStabilityStatusChanged;
 
+        public Action<bool, int, double, int> OnStabilityStatusChanged { get => m_OnStabilityStatusChanged; set => m_OnStabilityStatusChanged = value; }
         /// <summary>
         /// Set on true when SP deactivates boosting and enter the stable state.
         /// Once SP enters the stable state and it becomes instable again, this value is set on false.
@@ -105,7 +111,7 @@ namespace NeoCortexApi
         /// </summary>
         /// <param name="input">The input of the SP in the current cycle.</param>
         /// <param name="output">The output SDR of the Spatial Pooler compute cycle.</param>
-        /// <returns></returns>
+        /// <returns>True if the PS has enetered the stable state.</returns>
         public bool Compute(int[] input, int[] output)
         {
             bool res = false;
@@ -235,6 +241,10 @@ namespace NeoCortexApi
             return sum / max;
         }
 
+        public void SetConnections(Connections connections)
+        {
+            this.m_HtmMemory = connections;
+        }
 
         /// <summary>
         /// Calculates how many elements of the array are same in percents. This method is useful to compare 
@@ -285,6 +295,93 @@ namespace NeoCortexApi
                 return Encoding.UTF8.GetString(data);
             }
         }
+
+        //public static HomeostaticPlasticityController Deserialize(StreamReader sr, Connections htmMemory = null)
+        //{
+        //    HomeostaticPlasticityController ctrl = new HomeostaticPlasticityController();
+        //    ctrl.m_HtmMemory = htmMemory;
+
+        //    HtmSerializer2 ser = new HtmSerializer2();
+
+        //    while (sr.Peek() >= 0)
+        //    {
+        //        string data = sr.ReadLine();
+        //        if (data == String.Empty || data == ser.ReadBegin(nameof(HomeostaticPlasticityController)))
+        //        {
+        //            continue;
+        //        }
+        //        else if (data == ser.ReadBegin(nameof(Connections)))
+        //        {
+        //            ctrl.m_HtmMemory = Connections.Deserialize(sr);
+        //        }
+        //        else if (data == ser.ReadEnd(nameof(HomeostaticPlasticityController)))
+        //        {
+        //            break;
+        //        }
+        //        else
+        //        {
+        //            string[] str = data.Split(HtmSerializer2.ParameterDelimiter);
+        //            for (int i = 0; i < str.Length; i++)
+        //            {
+        //                switch (i)
+        //                {
+        //                    case 0:
+        //                        {
+        //                            ctrl.m_RequiredSimilarityThreshold = ser.ReadDoubleValue(str[i]);
+        //                            break;
+        //                        }
+        //                    case 1:
+        //                        {
+        //                            ctrl.m_MaxPreviousElements = ser.ReadIntValue(str[i]);
+        //                            break;
+        //                        }
+        //                    case 2:
+        //                        {
+        //                            ctrl.m_Cycle = ser.ReadIntValue(str[i]);
+        //                            break;
+        //                        }
+        //                    case 3:
+        //                        {
+        //                            ctrl.m_MinCycles = ser.ReadIntValue(str[i]);
+        //                            break;
+        //                        }
+        //                    case 4:
+        //                        {
+        //                            ctrl.m_RequiredNumOfStableCycles = ser.ReadIntValue(str[i]);
+        //                            break;
+        //                        }
+        //                    case 5:
+        //                        {
+        //                            ctrl.m_NumOfStableCyclesForInput = ser.ReadDictSIValue(str[i]);
+        //                            break;
+        //                        }
+        //                    case 6:
+        //                        {
+        //                            ctrl.m_NumOfActiveColsForInput = ser.ReadDictSIarray(str[i]);
+        //                            break;
+        //                        }
+        //                    case 7:
+        //                        {
+        //                            ctrl.m_InOutMap = ser.ReadDictSIarray(str[i]);
+        //                            break;
+        //                        }
+
+        //                    case 8:
+        //                        {
+        //                            ctrl.m_IsStable = ser.ReadBoolValue(str[i]);
+        //                            break;
+        //                        }
+        //                    default:
+        //                        { break; }
+
+        //                }
+        //            }
+        //        }
+        //    }
+
+        //    return ctrl;
+
+        //}
 
         /// <summary>
         /// Traces out all cell indicies grouped by input value.
@@ -363,7 +460,7 @@ namespace NeoCortexApi
                 return false;
             else if (m_IsStable != obj.m_IsStable)
                 return false;
-           
+
             return true;
 
         }
@@ -374,7 +471,7 @@ namespace NeoCortexApi
             HtmSerializer2 ser = new HtmSerializer2();
 
             ser.SerializeBegin(nameof(HomeostaticPlasticityController), writer);
-            
+
             ser.SerializeValue(this.m_RequiredSimilarityThreshold, writer);
             ser.SerializeValue(this.m_MaxPreviousElements, writer);
             ser.SerializeValue(this.m_Cycle, writer);
@@ -384,14 +481,14 @@ namespace NeoCortexApi
             ser.SerializeValue(this.m_NumOfActiveColsForInput, writer);
             ser.SerializeValue(this.m_InOutMap, writer);
             ser.SerializeValue(this.m_IsStable, writer);
-            
+
             if (this.m_HtmMemory != null)
             {
                 this.m_HtmMemory.Serialize(writer);
             }
-            
+
             ser.SerializeEnd(nameof(HomeostaticPlasticityController), writer);
-            
+
         }
 
         public static HomeostaticPlasticityController Deserialize(StreamReader sr)
@@ -462,7 +559,7 @@ namespace NeoCortexApi
                                     ctrl.m_InOutMap = ser.ReadDictSIarray(str[i]);
                                     break;
                                 }
-                            
+
                             case 8:
                                 {
                                     ctrl.m_IsStable = ser.ReadBoolValue(str[i]);
@@ -478,6 +575,54 @@ namespace NeoCortexApi
 
             return ctrl;
 
+        }
+
+        public void Serialize(object obj, string name, StreamWriter sw)
+        {
+            var excludeEntries = new List<string>
+            {
+                nameof(m_OnStabilityStatusChanged),
+                nameof(OnStabilityStatusChanged),
+                nameof(m_HtmMemory),
+                nameof(m_InOutMap)
+            };
+
+            if (obj is HomeostaticPlasticityController controller)
+            {
+                HtmSerializer2.SerializeObject(obj, name, sw, excludeEntries);
+                var convertInOutMap = controller.m_InOutMap.ToDictionary(kv => kv.Key, kv => new KeyValuePair<int, int[]>(kv.Value.Length, ArrayUtils.IndexesWithNonZeros(kv.Value)));
+                HtmSerializer2.Serialize(convertInOutMap, "convertInOutMap", sw);
+            }
+
+
+        }
+
+        public static object Deserialize<T>(StreamReader sr, string name)
+        {
+            var excludeEntries = new List<string> { "convertInOutMap" };
+
+            var controller = HtmSerializer2.DeserializeObject<T>(sr, name, excludeEntries, (obj, propName) =>
+            {
+                if (obj is HomeostaticPlasticityController hpc)
+                {
+                    if (propName == "convertInOutMap")
+                    {
+                        var convertInOutMap = HtmSerializer2.Deserialize<Dictionary<string, KeyValuePair<int, int[]>>>(sr, propName);
+
+                        Dictionary<string, int[]> inOutMap = new Dictionary<string, int[]>();
+                        foreach (var map in convertInOutMap)
+                        {
+                            var array = new int[map.Value.Key];
+                            ArrayUtils.FillArray(array, 0);
+                            ArrayUtils.SetIndexesTo(array, map.Value.Value, 1);
+                            inOutMap[map.Key] = array;
+                        }
+                        hpc.m_InOutMap = inOutMap;
+                    }
+                }
+            });
+
+            return controller;
         }
         #endregion
     }
