@@ -5,6 +5,7 @@ using NeoCortexApi.Utility;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
 
 namespace NeoCortexApi
@@ -12,7 +13,7 @@ namespace NeoCortexApi
     /// <summary>
     /// Implementation of Temporal Memory algorithm.
     /// </summary>
-    public class TemporalMemory : IHtmAlgorithm<int[], ComputeCycle>//: IComputeDecorator
+    public class TemporalMemory : IHtmAlgorithm<int[], ComputeCycle>, ISerializable//: IComputeDecorator
     {
         private static readonly double EPSILON = 0.00001;
 
@@ -43,11 +44,11 @@ namespace NeoCortexApi
         {
             this.connections = conn;
 
-            SparseObjectMatrix<Column> matrix = this.connections.HtmConfig.Memory == null ?
+            SparseObjectMatrix<Column> matrix = this.connections.Memory == null ?
                 new SparseObjectMatrix<Column>(this.connections.HtmConfig.ColumnDimensions) :
-                    (SparseObjectMatrix<Column>)this.connections.HtmConfig.Memory;
+                    (SparseObjectMatrix<Column>)this.connections.Memory;
 
-            this.connections.HtmConfig.Memory = matrix;
+            this.connections.Memory = matrix;
 
             int numColumns = matrix.GetMaxIndex() + 1;
             this.connections.HtmConfig.NumColumns = numColumns;
@@ -58,7 +59,9 @@ namespace NeoCortexApi
             Column colZero = matrix.GetObject(0);
             for (int i = 0; i < numColumns; i++)
             {
-                Column column = colZero == null ? new Column(cellsPerColumn, i, this.connections.HtmConfig.SynPermConnected, this.connections.HtmConfig.NumInputs) : matrix.GetObject(i);
+                Column column = colZero == null ? 
+                    new Column(cellsPerColumn, i, this.connections.HtmConfig.SynPermConnected, this.connections.HtmConfig.NumInputs) : matrix.GetObject(i);
+
                 for (int j = 0; j < cellsPerColumn; j++)
                 {
                     cells[i * cellsPerColumn + j] = column.Cells[j];
@@ -111,19 +114,20 @@ namespace NeoCortexApi
         /// <remarks>Note: PredictiveCells are not calculated here. They are calculated on demand from active segments.</remarks>
         public ComputeCycle Compute(int[] activeColumns, bool learn, int[] externalPredictiveInputsActive = null, int[] externalPredictiveInputsWinners = null)
         {
-            Stopwatch sw = new Stopwatch();
-            sw.Start();
+            //Stopwatch sw = new Stopwatch();
+            //sw.Start();
+
             ComputeCycle cycle = ActivateCells(this.connections, activeColumns, learn);
 
-            sw.Stop();
+            //sw.Stop();
 
-            var sec1 = sw.ElapsedMilliseconds;
+            //var sec1 = sw.ElapsedMilliseconds;
 
-            sw.Restart();
+            //sw.Restart();
 
             ActivateDendrites(this.connections, cycle, learn, externalPredictiveInputsActive, externalPredictiveInputsWinners);
 
-            sw.Stop();
+            //sw.Stop();
 
             //tmperf1.WriteLine($"{sec1}, {sw.ElapsedMilliseconds}, {sec1 + sw.ElapsedMilliseconds}");
 
@@ -199,7 +203,7 @@ namespace NeoCortexApi
             Func<Object, Column> segToCol = (segment) =>
             {
                 var colIndx = ((DistalDendrite)segment).ParentCell.ParentColumnIndex;
-                var parentCol = this.connections.HtmConfig.Memory.GetColumn(colIndx);
+                var parentCol = this.connections.Memory.GetColumn(colIndx);
                 return parentCol;
             };
 
@@ -226,7 +230,7 @@ namespace NeoCortexApi
                     // If there are some active segments on the column already...
                     if (activeColumnData.ActiveSegments != null && activeColumnData.ActiveSegments.Count > 0)
                     {
-                        //Debug.Write("A");
+                        //Debug.Write("A");1    
 
                         List<Cell> cellsOwnersOfActSegs = ActivatePredictedColumn(conn, activeColumnData.ActiveSegments,
                             activeColumnData.MatchingSegments, prevActiveCells, prevWinnerCells,
@@ -247,25 +251,14 @@ namespace NeoCortexApi
                             prevActiveCells, prevWinnerCells, permanenceIncrement, permanenceDecrement, conn.HtmConfig.Random,
                                learn);
 
-                        // DRAFT. Removing this as unnecessary.
-                        //cycle.ActiveCells.Add(burstingResult.BestCell);
-
-
                         // Here we activate all cells by putting them to list of active cells.
                         cycle.ActiveCells.AddRange(burstingResult.Cells);
 
-                        //foreach (var item in burstingResult.Cells)
-                        //{
-                        //    cycle.ActiveCells.Add(item);
-                        //}
+                        // Test was done. Better performance is when BestCell is used only instead of adding all cells.
+                        //cycle.WinnerCells.AddRange(burstingResult.Cells);
 
-                        //var actSyns = conn.getReceptorSynapses(burstingResult.BestCell).Where(s=>prevActiveCells.Contains(s.SourceCell));
-                        //foreach (var syn in actSyns)
-                        //{
-                        //    cycle.ActiveSynapses.Add(syn);
-                        //}
-
-                        cycle.WinnerCells.Add((Cell)burstingResult.BestCell);
+                        // The winner cell is added to th elots of winner cells in the cycle.
+                        cycle.WinnerCells.Add(burstingResult.BestCell);
                     }
                 }
                 else
@@ -318,9 +311,12 @@ namespace NeoCortexApi
             var activeSegments = new List<DistalDendrite>();
             foreach (var item in activity.ActiveSynapses)
             {
-                var seg = conn.GetSegmentForFlatIdx(item.Key);
-                if (seg != null && item.Value >= conn.HtmConfig.ActivationThreshold)
-                    activeSegments.Add(seg);
+                if (item.Value >= conn.HtmConfig.ActivationThreshold)
+                {
+                    var seg = conn.GetSegmentForFlatIdx(item.Key);
+                    if (seg != null)
+                        activeSegments.Add(seg);
+                }
             }
 
             //
@@ -564,6 +560,8 @@ namespace NeoCortexApi
                 leastUsedOrMaxPotentialCell = GetLeastUsedCell(conn, cells, random);
                 if (learn)
                 {
+                    // This can be optimized. Right now, we assume that every winner cell has a single synaptic connection to the segment.
+                    // This is why we substract number of cells from the MaxNewSynapseCount.
                     int nGrowExact = Math.Min(conn.HtmConfig.MaxNewSynapseCount, prevWinnerCells.Count);
                     if (nGrowExact > 0)
                     {
@@ -610,7 +608,7 @@ namespace NeoCortexApi
         }
 
         /// <summary>
-        /// Punishes the Segments that incorrectly predicted a column to be active.
+        /// Punishes the MatchingSegments that incorrectly predicted a column to be active.
         /// <para>
         /// Pseudocode:<br/>
         ///  for each matching segment in the column<br/>
@@ -619,7 +617,7 @@ namespace NeoCortexApi
         /// </summary>
         /// <param name="conn">Connections instance for the <see cref="TemporalMemory"/></param>
         /// <param name="activeSegments">An iterable of <see cref="DistalDendrite"/> actives</param>
-        /// <param name="matchingSegments">An iterable of <see cref="DistalDendrite"/> matching for the column compute is operating on that are matching; None if empty</param>
+        /// <param name="matchingSegments"><see cref="DistalDendrite"/> matching for the column.</param>
         /// <param name="prevActiveCells">Active cells in `t-1`</param>
         /// <param name="prevWinnerCells">Winner cells in `t-1` are decremented during learning.</param>
         /// <param name="predictedSegmentDecrement">Amount by which segments are punished for incorrect predictions</param>
@@ -634,14 +632,15 @@ namespace NeoCortexApi
                 {
                     AdaptSegment(conn, segment, prevActiveCells, -conn.HtmConfig.PredictedSegmentDecrement, 0);
                 }
+
+                //foreach (DistalDendrite segment in activeSegments)
+                //{
+                //    AdaptSegment(conn, segment, prevActiveCells, -conn.HtmConfig.PredictedSegmentDecrement, 0);
+                //}
             }
         }
 
         #region Helper Methods
-        ////////////////////////////
-        //     Helper Methods     //
-        ////////////////////////////
-
 
         /// <summary>
         /// Gets the cell with the smallest number of segments in the currentlly processing mini-column.
@@ -698,8 +697,9 @@ namespace NeoCortexApi
             //
             // Enumarates all synapses in a segment and remove winner-cells from
             // list of removingCandidates if they are presynaptic winners cells.
-            // So, we will create synapses only on cells, which are not winners in the previous cycle.
-            //DD 
+            // So, we will create synapses only from cells, which do not already have synaptic connection to the segment. 
+            // DD: Removing of this loop creates same result as with the loop.
+            // This should be investigated.
             foreach (Synapse synapse in segment.Synapses)
             {
                 Cell presynapticCell = synapse.GetPresynapticCell();
@@ -851,6 +851,52 @@ namespace NeoCortexApi
         public DentriteComparer GetComparer(int nextSegmentOrdinal)
         {
             return new DentriteComparer(nextSegmentOrdinal);
+        }
+
+        public bool Equals(TemporalMemory obj)
+        {
+            if (this == obj)
+                return true;
+            if (obj == null)
+                return false;
+
+            if (this.connections == null)
+            {
+                if (obj.connections != null)
+                    return false;
+            }
+            else if (this.indxOfLastHighestSegment != obj.indxOfLastHighestSegment)
+                return false;
+            else if (this.LastActivity == null)
+            {
+                if (obj.LastActivity != null)
+                    return false;
+            }
+            else if (!this.LastActivity.Equals(obj.LastActivity))
+                return false;
+
+            return true;
+        }
+
+        public bool Equals(IHtmModule other)
+        {
+            if (!(other is TemporalMemory tm))
+                return false;
+            return this.Equals(tm);
+        }
+
+        public void Serialize(object obj, string name, StreamWriter sw)
+        {
+            var ignoreMembers = new List<string>
+            {
+                //nameof(TemporalMemory.connections)
+            };
+            HtmSerializer2.SerializeObject(obj, name, sw, ignoreMembers);  
+        }
+
+        public static object Deserialize<T>(StreamReader sr, string name)
+        {
+            return HtmSerializer2.DeserializeObject<T>(sr, name);
         }
     }
 }
