@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 using System;
 using System.Collections.Generic;
+using System.IO;
 
 namespace NeoCortexApi.Entities
 {
@@ -9,7 +10,7 @@ namespace NeoCortexApi.Entities
     /// <summary>
     /// Provides common generic independent calculation functions.
     /// </summary>
-    public class AbstractFlatMatrix
+    public class AbstractFlatMatrix /*: ISerializable*/
     {
         public AbstractFlatMatrix()
         {
@@ -118,6 +119,16 @@ namespace NeoCortexApi.Entities
             }
             return dimensionMultiples;
         }
+
+        //public void Serialize(object obj, string name, StreamWriter sw)
+        //{
+        //    HtmSerializer2.SerializeObject(obj, name, sw);
+        //}
+
+        //public static object Deserialize(StreamReader sr, string name)
+        //{
+        //    return HtmSerializer2.DeserializeObject<AbstractFlatMatrix>(sr, name);
+        //}
     }
 
 
@@ -342,7 +353,7 @@ namespace NeoCortexApi.Entities
          */
         //@SuppressWarnings("rawtypes")
         //@Override
-        public bool Equals(AbstractFlatMatrix<T> obj)
+        public virtual bool Equals(AbstractFlatMatrix<T> obj)
         {
             if (this == obj)
                 return true;
