@@ -11,7 +11,7 @@ namespace NeoCortexApi.Entities
     /// <summary>
     /// Defines a single cell (neuron).
     /// </summary>
-    public class Cell : IEquatable<Cell>, IComparable<Cell>
+    public class Cell : IEquatable<Cell>, IComparable<Cell>, ISerializable
     {
 
         /// <summary>
@@ -22,7 +22,7 @@ namespace NeoCortexApi.Entities
         /// <summary>
         /// The identifier of the cell.
         /// </summary>
-        public int CellId { get; private set; }
+        //public int CellId { get; private set; }
 
 
         /// <summary>
@@ -36,12 +36,17 @@ namespace NeoCortexApi.Entities
         private readonly int m_Hashcode;
 
         /// <summary>
-        /// List of dendrites of the cell. Every dendrite segment is owned bt the cell.
+        /// List of dendrites of the cell.
         /// </summary>
         public List<DistalDendrite> DistalDendrites { get; set; } = new List<DistalDendrite>();
 
         /// <summary>
-        /// List of receptor synapses that connect this cells as a source (presynaptic) cell to the distal dendrite segment owned by some other cell.
+        /// List of apical dendrites of the cell. 
+        /// </summary>
+        public List<ApicalDendrite> ApicalDendrites { get; set; } = new List<ApicalDendrite>();
+
+        /// <summary>
+        /// List of receptor synapses (outgoing synapses) that connect this cells as a source (presynaptic) cell to the distal dendrite segment owned by some other cell.
         /// This synapse assotiates the cell (presynaptic = source cell) with the next cell (postsynaptic = destination cell).
         /// The destination cell is the parent cell of the segment to which the presynaptic cell is connected.
         /// </summary>
@@ -63,13 +68,13 @@ namespace NeoCortexApi.Entities
         /// <param name="numCellsPerColumn"></param>
         /// <param name="cellId"></param>
         /// <param name="cellActivity"></param>
-        public Cell(int parentColumnIndx, int colSeq, int numCellsPerColumn, int cellId, CellActivity cellActivity)
+        public Cell(int parentColumnIndx, int colSeq, int numCellsPerColumn, CellActivity cellActivity)
         {
             this.ParentColumnIndex = parentColumnIndx;
 
             this.Index = parentColumnIndx * numCellsPerColumn + colSeq;
 
-            this.CellId = cellId;
+            //this.CellId = cellId;
         }
 
         /// <summary>
@@ -78,16 +83,21 @@ namespace NeoCortexApi.Entities
         /// <returns></returns>
         public override int GetHashCode()
         {
-            if (m_Hashcode == 0)
-            {
-                int prime = 31;
-                int result = 1;
-                result = prime * result + Index;
-                return result;
-            }
-            return m_Hashcode;
+            int prime = 31;
+            int result = 1;
+            result = prime * result + ParentColumnIndex;
+            result = prime * result + Index;
+            return result;
+            //if (m_Hashcode == 0)
+            //{
+            //    int prime = 31;
+            //    int result = 1;
+            //    result = prime * result + ParentColumnIndex;
+            //    result = prime * result + Index;
+            //    return result;
+            //}
+            //return m_Hashcode;
         }
-
 
         /// <summary>
         /// <inheritdoc/>
@@ -108,20 +118,20 @@ namespace NeoCortexApi.Entities
                 if (obj.ParentColumnIndex != this.ParentColumnIndex)
                     return false;
 
-                if (obj.CellId != this.CellId)
-                    return false;
+                //if (obj.CellId != this.CellId)
+                //    return false;
 
                 if (obj.DistalDendrites != null && this.DistalDendrites != null)
                 {
                     //if (!Enumerable.SequenceEqual(obj.DistalDendrites, this.DistalDendrites))
                     //    return false;
-                    if (!obj.DistalDendrites.SequenceEqual(this.DistalDendrites))
+                    if (!obj.DistalDendrites.ElementsEqual(this.DistalDendrites))
                         return false;
                 }
 
                 if (obj.ReceptorSynapses != null && this.ReceptorSynapses != null)
                 {
-                    if (!obj.ReceptorSynapses.SequenceEqual(this.ReceptorSynapses))
+                    if (!obj.ReceptorSynapses.ElementsEqual(this.ReceptorSynapses))
                         return false;
                 }
 
@@ -160,14 +170,14 @@ namespace NeoCortexApi.Entities
         /// Serializes the cell to the stream.
         /// </summary>
         /// <param name="writer"></param>
-        public void  SerializeT(StreamWriter writer)
+        public void SerializeT(StreamWriter writer)
         {
             HtmSerializer2 ser = new HtmSerializer2();
 
             ser.SerializeBegin(nameof(Cell), writer);
 
             ser.SerializeValue(this.Index, writer);
-            ser.SerializeValue(this.CellId, writer);
+            //ser.SerializeValue(this.CellId, writer);
             ser.SerializeValue(this.ParentColumnIndex, writer);
 
             if (this.DistalDendrites != null && this.DistalDendrites.Count > 0)
@@ -178,7 +188,7 @@ namespace NeoCortexApi.Entities
 
             ser.SerializeEnd(nameof(Cell), writer);
 
-            
+
         }
 
         public static Cell Deserialize(StreamReader sr)
@@ -196,7 +206,7 @@ namespace NeoCortexApi.Entities
                 }
                 else if (data == ser.ReadBegin(nameof(DistalDendrite)))
                 {
-                    cell.DistalDendrites.Add(DistalDendrite.Deserialize(sr));
+                    //cell.DistalDendrites.Add(DistalDendrite.Deserialize(sr));
                 }
                 else if (data == ser.ReadBegin(nameof(Synapse)))
                 {
@@ -220,7 +230,7 @@ namespace NeoCortexApi.Entities
                                 }
                             case 1:
                                 {
-                                    cell.CellId = ser.ReadIntValue(str[i]);
+                                    // cell.CellId = ser.ReadIntValue(str[i]);
                                     break;
                                 }
                             case 2:
@@ -235,6 +245,34 @@ namespace NeoCortexApi.Entities
                     }
                 }
             }
+            return cell;
+        }
+
+        public void Serialize(object obj, string name, StreamWriter sw)
+        {
+            var ignoreMembers = new List<string> 
+            { 
+                nameof(Cell.ReceptorSynapses),
+                nameof(m_Hashcode)
+            };
+            HtmSerializer2.SerializeObject(obj, name, sw, ignoreMembers);
+            var cell = obj as Cell;
+            if (cell != null)
+            {
+                var synapses = cell.ReceptorSynapses.Select(s => new Synapse() { SynapseIndex = s.SynapseIndex });
+                HtmSerializer2.Serialize(synapses, nameof(Cell.ReceptorSynapses), sw);
+            }
+        }
+        public static object Deserialize<T>(StreamReader sr, string name)
+        {
+            if (typeof(T) != typeof(Cell))
+                return null;
+            var cell = HtmSerializer2.DeserializeObject<Cell>(sr, name);
+
+            //foreach (var distalDentrite in cell.DistalDendrites)
+            //{
+            //    distalDentrite.ParentCell = cell;
+            //}
             return cell;
         }
         #endregion
