@@ -1,9 +1,11 @@
 ﻿// Copyright (c) Damir Dobric. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net.Http.Headers;
+using System.Reflection;
 using System.Text;
 
 namespace NeoCortexApi.Entities
@@ -12,7 +14,7 @@ namespace NeoCortexApi.Entities
     /// HTM configuration.
     /// Also sent from Akka-client to Akka Actor.
     /// </summary>
-    public class HtmConfig
+    public class HtmConfig : ISerializable
     {
         /// <summary>
         /// Default constructor with the default set of parameters.
@@ -45,7 +47,7 @@ namespace NeoCortexApi.Entities
 
         private double synPermActiveInc;
         private double synPermConnected;
-        private AbstractSparseMatrix<Column> memory;
+        
         private ISparseMatrix<int> inputMatrix;
 
         public TemporalMemoryConfig TemporalMemory { get; set; } = new TemporalMemoryConfig();
@@ -319,11 +321,6 @@ namespace NeoCortexApi.Entities
         public HtmModuleTopology InputModuleTopology { get; set; }
 
         /// <summary>
-        /// The main data structure containing columns, cells, and synapses.
-        /// </summary>
-        public AbstractSparseMatrix<Column> Memory { get => memory; set { memory = value; ColumnModuleTopology = value?.ModuleTopology; } }
-
-        /// <summary>
         /// Activation threshold used in sequence learning. If the number of active connected synapses on a distal segment is at least this threshold, the segment is declared as active one.
         /// </summary>
         public int ActivationThreshold { get; set; } = 13;
@@ -417,7 +414,6 @@ namespace NeoCortexApi.Entities
 
         public void ClearModuleTopology()
         {
-            this.Memory = null;
             this.InputMatrix = null;
         }
         public bool Equals(HtmConfig obj)
@@ -427,13 +423,13 @@ namespace NeoCortexApi.Entities
             if (obj == null)
                 return false;
 
-            if (memory == null)
-            {
-                if (obj.memory != null)
-                    return false;
-            }
-            else if (!memory.Equals(obj.memory))
-                return false;
+            //if (memory == null)
+            //{
+            //    if (obj.memory != null)
+            //        return false;
+            //}
+            //else if (!memory.Equals(obj.memory))
+            //    return false;
             if (inputMatrix == null)
             {
                 if (obj.inputMatrix != null)
@@ -476,13 +472,13 @@ namespace NeoCortexApi.Entities
             }
             else if (!InputModuleTopology.Equals(obj.InputModuleTopology))
                 return false;
-            if (Memory == null)
-            {
-                if (obj.Memory != null)
-                    return false;
-            }
-            else if (!Memory.Equals(obj.Memory))
-                return false;
+            //if (Memory == null)
+            //{
+            //    if (obj.Memory != null)
+            //        return false;
+            //}
+            //else if (!Memory.Equals(obj.Memory))
+            //    return false;
             if (synPermActiveInc != obj.SynPermActiveInc)
                 return false;
             if (synPermConnected != obj.synPermConnected)
@@ -533,7 +529,7 @@ namespace NeoCortexApi.Entities
                 return false;
             if (MaxBoost != obj.MaxBoost)
                 return false;
-           
+
             if (UpdatePeriod != obj.UpdatePeriod)
                 return false;
             if (OverlapDutyCycles != null && this.OverlapDutyCycles != null)
@@ -600,8 +596,7 @@ namespace NeoCortexApi.Entities
                 return false;
             if (Random != null && obj.Random != null)
             {
-
-                if (obj.Random.Next() != Random.Next())
+                if (!obj.Random.Equals(Random))
                     return false;
             }
 
@@ -619,6 +614,7 @@ namespace NeoCortexApi.Entities
             ser.SerializeValue(this.SynPermConnected, writer);
             //Spatial Pooler Variables
             //ser.SerializeValue(this.InhibitionRadius, writer);
+            ser.SerializeValue(-1, writer);
             ser.SerializeValue(this.NumInputs, writer);
             ser.SerializeValue(this.NumColumns, writer);
             ser.SerializeValue(this.PotentialRadius, writer);
@@ -642,7 +638,7 @@ namespace NeoCortexApi.Entities
             ser.SerializeValue(this.PredictedSegmentDecrement, writer);
             ser.SerializeValue(this.DutyCyclePeriod, writer);
             ser.SerializeValue(this.MaxBoost, writer);
-           
+
             ser.SerializeValue(this.UpdatePeriod, writer);
             ser.SerializeValue(this.OverlapDutyCycles, writer);
             ser.SerializeValue(this.ActiveDutyCycles, writer);
@@ -666,10 +662,10 @@ namespace NeoCortexApi.Entities
             ser.SerializeValue(this.Name, writer);
             ser.SerializeValue(this.RandomGenSeed, writer);
 
-            if (this.memory != null)
-            {
-                this.memory.Serialize(writer);
-            }
+            //if (this.memory != null)
+            //{
+            //    this.memory.Serialize(writer);
+            //}
             if (this.inputMatrix != null)
             {
                 this.inputMatrix.Serialize(writer);
@@ -695,10 +691,10 @@ namespace NeoCortexApi.Entities
             {
                 this.InputMatrix.Serialize(writer);
             }
-            if (this.Memory != null)
-            {
-                this.Memory.Serialize(writer);
-            }
+            //if (this.Memory != null)
+            //{
+            //    this.Memory.Serialize(writer);
+            //}
 
             ser.SerializeEnd(nameof(HtmConfig), writer);
         }
@@ -998,6 +994,23 @@ namespace NeoCortexApi.Entities
 
                 }
             }
+            return htmConfig;
+        }
+
+        public void Serialize(object obj, string name, StreamWriter sw)
+        {
+            var excludeMembers = new List<string>
+            {
+                nameof(HtmConfig.inputMatrix),
+                nameof(HtmConfig.synPermActiveInc),
+                nameof(HtmConfig.synPermConnected)
+            };
+            HtmSerializer2.SerializeObject(obj, name, sw, excludeMembers);
+        }
+
+        public static object Deserialize<T>(StreamReader sr, string name)
+        {
+            var htmConfig = HtmSerializer2.DeserializeObject<HtmConfig>(sr, name);
             return htmConfig;
         }
         #endregion
