@@ -3,6 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 
 namespace NeoCortexApi.Entities
@@ -99,9 +100,9 @@ namespace NeoCortexApi.Entities
                 return true;
             if (obj == null)
                 return false;
-            if (dimensions != obj.dimensions)
+            if (!dimensions.SequenceEqual(obj.dimensions))
                 return false;
-            if (dimensionMultiples != obj.dimensionMultiples)
+            if (!dimensionMultiples.SequenceEqual(obj.dimensionMultiples))
                 return false;
             if (isColumnMajor != obj.isColumnMajor)
                 return false;
