@@ -9,7 +9,7 @@ using System.IO;
 
 namespace NeoCortexApi.Entities
 {
-    public class InMemoryArray : IDistributedArray
+    public class InMemoryArray : IDistributedArray, ISerializable
     {
         public int[] dimensions;
 
@@ -201,6 +201,14 @@ namespace NeoCortexApi.Entities
                 backingArray.SetValue(0, rowIndex, i);
             }
         }
+        public override bool Equals(object obj)
+        {
+            var array = obj as InMemoryArray;
+            if (array == null)
+                return false;
+            return this.Equals(array);
+        }
+
         public bool Equals(InMemoryArray obj)
         {
             if (this == obj)
@@ -208,15 +216,53 @@ namespace NeoCortexApi.Entities
 
             if (obj == null)
                 return false;
-            if (this.dimensions != obj.dimensions)
-                return false;
-            else if (this.backingArray != obj.backingArray)
+            //if (!this.dimensions.SequenceEqual(obj.dimensions))
+            //    return false;
+            else if (!ArrayEquals(this.backingArray, obj.backingArray))
                 return false;
             else if (this.numOfNodes != obj.numOfNodes)
                 return false;
             else if (this.Rank != obj.Rank)
                 return false;
-            else if (this.Dimensions != obj.Dimensions)
+            else if (!this.Dimensions.SequenceEqual(obj.Dimensions))
+                return false;
+
+            return true;
+        }
+
+        private bool ArrayEquals(Array array1, Array array2)
+        {
+            if (array1 == null)
+                return array2 == null;
+            if (array2 == null)
+                return false;
+
+            if (array1.Rank != array2.Rank)
+                return false;
+            for (int r = 0; r < array1.Rank; r++)
+            {
+                if (array1.GetLength(r) != array2.GetLength(r))
+                    return false;
+            }
+
+            var arrayType = array1.GetType();
+            var elementType = arrayType.GetElementType();
+
+            var castMethod = typeof(Enumerable).GetMethod(nameof(Enumerable.Cast))?.MakeGenericMethod(elementType);
+            if (castMethod == null)
+            {
+                throw new Exception("No cast method. This should not happen!");
+            }
+
+            var list1 = castMethod.Invoke(null, new object[] { array1 });
+            var list2 = castMethod.Invoke(null, new object[] { array2 });
+
+
+            var sequenceEqualMethod = typeof(Enumerable).GetMethods().FirstOrDefault(m => m.Name == nameof(Enumerable.SequenceEqual) && m.GetParameters().Length == 2)?.MakeGenericMethod(elementType);
+
+            var isSequenceEqual = (bool)sequenceEqualMethod.Invoke(null, new object[] { list1, list2 });
+
+            if (!isSequenceEqual)
                 return false;
 
             return true;
@@ -227,7 +273,7 @@ namespace NeoCortexApi.Entities
             HtmSerializer2 ser = new HtmSerializer2();
 
             ser.SerializeBegin(nameof(InMemoryArray), writer);
-            
+
             ser.SerializeValue(this.backingArray, writer);
             ser.SerializeValue(this.dimensions, writer);
             ser.SerializeValue(this.numOfNodes, writer);
@@ -277,6 +323,24 @@ namespace NeoCortexApi.Entities
             }
 
             return array;
+        }
+
+        public void Serialize(object obj, string name, StreamWriter sw)
+        {
+            var ignoreMembers = new List<string> 
+            { 
+                "Item",
+                nameof(Dimensions),
+                nameof(IDistributedArray.Count),
+                nameof(Rank)
+            };
+            HtmSerializer2.SerializeObject(obj, name, sw, ignoreMembers);
+        }
+
+        public static object Deserialize<T>(StreamReader sr, string name)
+        {
+            var ignoreMembers = new List<string> { "Item" };
+            return HtmSerializer2.DeserializeObject<T>(sr, name, ignoreMembers);
         }
         #endregion
     }
