@@ -137,7 +137,7 @@ namespace UnitTestsProject
             tm.Init(cn);
 
             int[] activeColumns = { 0 };
-            IList<Cell> burstingCells = cn.GetCellSet(new int[] { 0, 1, 2, 3 });
+            var burstingCells = cn.GetCells(new int[] { 0, 1, 2, 3 });
 
             ComputeCycle cc = tm.Compute(activeColumns, true) as ComputeCycle;
 
@@ -155,7 +155,7 @@ namespace UnitTestsProject
             tm.Init(cn);
 
             int[] activeColumns = { 0 };
-            IList<Cell> burstingCells = cn.GetCellSet(new int[] { 0, 1, 2, 3 });
+            var burstingCells = cn.GetCells(new int[] { 0, 1, 2, 3 });
 
             ComputeCycle cc = tm.Compute(activeColumns, true) as ComputeCycle;
 
@@ -207,7 +207,7 @@ namespace UnitTestsProject
             int[] previousActiveColumns = { 0 };
             int[] activeColumns = { 1 };
             Cell[] previousActiveCells = { cn.GetCell(0), cn.GetCell(1), cn.GetCell(2), cn.GetCell(3) };
-            List<Cell> expectedWinnerCells = new List<Cell>(cn.GetCellSet(new int[] { 4, 6 }));
+            List<Cell> expectedWinnerCells = new List<Cell>(cn.GetCells(new int[] { 4, 6 }));
 
             DistalDendrite activeSegment1 = cn.CreateDistalSegment(expectedWinnerCells[0]);
             cn.CreateSynapse(activeSegment1, previousActiveCells[0], 0.5);
@@ -482,7 +482,7 @@ namespace UnitTestsProject
             tm.Init(cn);
 
             int[] previousActiveColumns = { 0, 1, 2, 3 };
-            IList<Cell> prevWinnerCells = cn.GetCellSet(new int[] { 0, 1, 2, 3 });
+            IList<Cell> prevWinnerCells = cn.GetCells(new int[] { 0, 1, 2, 3 });
             int[] activeColumns = { 4 };
 
             DistalDendrite matchingSegment = cn.CreateDistalSegment(cn.GetCell(4));
@@ -522,7 +522,7 @@ namespace UnitTestsProject
             tm.Init(cn);
 
             int[] previousActiveColumns = { 0, 1 };
-            IList<Cell> prevWinnerCells = cn.GetCellSet(new int[] { 0, 1 });
+            IList<Cell> prevWinnerCells = cn.GetCells(new int[] { 0, 1 });
             int[] activeColumns = { 4 };
 
             DistalDendrite matchingSegment = cn.CreateDistalSegment(cn.GetCell(4));
@@ -673,7 +673,7 @@ namespace UnitTestsProject
             Assert.AreEqual(3, cn.HtmConfig.MaxSynapsesPerSegment);
 
             int[] prevActiveColumns = { 0, 1, 2 };
-            IList<Cell> prevWinnerCells = cn.GetCellSet(new int[] { 0, 1, 2 });
+            IList<Cell> prevWinnerCells = cn.GetCells(new int[] { 0, 1, 2 });
             int[] activeColumns = { 4 };
 
             DistalDendrite matchingSegment = cn.CreateDistalSegment(cn.GetCell(4));
@@ -862,7 +862,7 @@ namespace UnitTestsProject
                 Cell[] prevActiveCells = { cn.GetCell(4), cn.GetCell(5), cn.GetCell(6), cn.GetCell(7) };
                 int[] activeColumns = { 0 };
                 Cell[] nonMatchingCells = { cn.GetCell(0), cn.GetCell(3) };
-                IList<Cell> activeCells = cn.GetCellSet(new int[] { 0, 1, 2, 3 });
+                IList<Cell> activeCells = cn.GetCells(new int[] { 0, 1, 2, 3 });
 
                 DistalDendrite segment1 = cn.CreateDistalSegment(nonMatchingCells[0]);
                 cn.CreateSynapse(segment1, prevActiveCells[0], 0.5);
@@ -908,7 +908,7 @@ namespace UnitTestsProject
                     Assert.AreEqual(0.2, synapse.Permanence, 0.01);
 
                     var parentColIndx = synapse.GetPresynapticCell().ParentColumnIndex;
-                    Column column = cn.HtmConfig.Memory.GetColumn(parentColIndx);
+                    Column column = cn.Memory.GetColumn(parentColIndx);
                     Assert.IsTrue(columnCheckList.Contains(column));
                     columnCheckList.Remove(column);
                 }
@@ -1035,7 +1035,7 @@ namespace UnitTestsProject
             Synapse s2 = cn.CreateSynapse(dd, cn.GetCell(37), 0.4);
             Synapse s3 = cn.CreateSynapse(dd, cn.GetCell(477), 0.9);
 
-            TemporalMemory.AdaptSegment(cn, dd, cn.GetCellSet(new int[] { 23, 37 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
+            TemporalMemory.AdaptSegment(cn, dd, cn.GetCells(new int[] { 23, 37 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
 
             Assert.AreEqual(0.7, s1.Permanence, 0.01);
             Assert.AreEqual(0.5, s2.Permanence, 0.01);
@@ -1055,11 +1055,11 @@ namespace UnitTestsProject
             DistalDendrite dd = cn.CreateDistalSegment(cn.GetCell(0));
             Synapse s1 = cn.CreateSynapse(dd, cn.GetCell(23), 0.9);
 
-            TemporalMemory.AdaptSegment(cn, dd, cn.GetCellSet(new int[] { 23 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
+            TemporalMemory.AdaptSegment(cn, dd, cn.GetCells(new int[] { 23 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
             Assert.AreEqual(1.0, s1.Permanence, 0.1);
 
             // Now permanence should be at max
-            TemporalMemory.AdaptSegment(cn, dd, cn.GetCellSet(new int[] { 23 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
+            TemporalMemory.AdaptSegment(cn, dd, cn.GetCells(new int[] { 23 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
             Assert.AreEqual(1.0, s1.Permanence, 0.1);
         }
 
@@ -1077,7 +1077,7 @@ namespace UnitTestsProject
             Synapse s1 = cn.CreateSynapse(dd, cn.GetCell(23), 0.1);
             cn.CreateSynapse(dd, cn.GetCell(1), 0.3);
 
-            TemporalMemory.AdaptSegment(cn, dd, cn.GetCellSet(new int[] { }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
+            TemporalMemory.AdaptSegment(cn, dd, cn.GetCells(new int[] { }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
             //DD Assert.IsFalse(cn.GetSynapses(dd).Contains(s1));
             Assert.IsFalse(dd.Synapses.Contains(s1));
         }
@@ -1109,7 +1109,13 @@ namespace UnitTestsProject
             p.apply(cn);
             tm.Init(cn);
 
-            Assert.AreEqual(64 * 64 * 32, cn.Cells.Length);
+            int cnt = 0;
+            foreach (var item in cn.GetColumns())
+            {
+                cnt+=item.Cells.Length;
+            }
+
+            Assert.AreEqual(64 * 64 * 32, cnt);
         }
 
         public void TemporalMemoryInit()
@@ -1121,6 +1127,204 @@ namespace UnitTestsProject
             TemporalMemory temporalMemory = new TemporalMemory();
 
             temporalMemory.Init(connections);
+        }
+
+        [TestMethod]
+        [TestCategory("Prod")]
+        public void TestWithTwoActiveColumns()
+        {
+            TemporalMemory tm = new TemporalMemory();
+            Connections cn = new Connections();
+            Parameters p = getDefaultParameters();
+            p.apply(cn);
+            tm.Init(cn);
+
+            int[] previousActiveColumns = { 2, 3 }; ///2 pre active cells
+            Cell cell5 = cn.GetCell(6); /// get cell 6 by calling connection method
+            Cell cell6 = cn.GetCell(7);
+
+            DistalDendrite activeSegment = cn.CreateDistalSegment(cell5);
+            //  DistalDendrite activeSegment1 = cn.CreateDistalSegment(cell6);
+            cn.CreateSynapse(activeSegment, cn.GetCell(0), 0.5);
+            cn.CreateSynapse(activeSegment, cn.GetCell(1), 0.5);
+            cn.CreateSynapse(activeSegment, cn.GetCell(2), 0.5);
+            cn.CreateSynapse(activeSegment, cn.GetCell(3), 0.5);
+
+
+            ComputeCycle cc = tm.Compute(previousActiveColumns, true) as ComputeCycle;
+            Assert.IsFalse(cc.ActiveCells.Count == 0);
+            Assert.IsFalse(cc.WinnerCells.Count == 0);
+            Assert.IsTrue(cc.PredictiveCells.Count == 0);
+
+            int[] zeroColumns = new int[0];
+            ComputeCycle cc2 = tm.Compute(zeroColumns, true) as ComputeCycle; ///learn = true
+            Assert.IsTrue(cc2.ActiveCells.Count == 0); /// Active cell ==0
+            Assert.IsTrue(cc2.WinnerCells.Count == 0);  /// wineer cell equal to 0
+            Assert.IsTrue(cc2.PredictiveCells.Count == 0); ///lost of depolirized cells equal to 0
+        }
+
+        ///<summary>
+        /// Test adapt segment from syapse to centre 
+        /// <Summary>
+        [TestMethod]
+        [TestCategory("Prod")]
+        public void TestAdaptSegmentToCentre()
+        {
+            TemporalMemory tm = new TemporalMemory();
+            Connections cn = new Connections();
+            Parameters p = Parameters.getAllDefaultParameters();
+            p.apply(cn);
+            tm.Init(cn);
+
+            DistalDendrite dd = cn.CreateDistalSegment(cn.GetCell(0));
+            Synapse s1 = cn.CreateSynapse(dd, cn.GetCell(23), 0.5); // central 
+
+            TemporalMemory.AdaptSegment(cn, dd, cn.GetCells(new int[] { 23 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
+            Assert.AreEqual(0.6, s1.Permanence, 0.1);
+
+            // Now permanence should be at mean
+            TemporalMemory.AdaptSegment(cn, dd, cn.GetCells(new int[] { 23 }), cn.HtmConfig.PermanenceIncrement, cn.HtmConfig.PermanenceDecrement);
+            Assert.AreEqual(0.7, s1.Permanence, 0.1);
+        }
+
+
+        /// <summary>
+        ///Test an Array which has none cells in it
+        /// </summary>
+        [TestMethod]
+        public void TestArrayNotContainingCells()
+        {
+
+            HtmConfig htmConfig = GetDefaultTMParameters();
+            Connections cn = new Connections(htmConfig);
+
+            TemporalMemory tm = new TemporalMemory();
+
+            tm.Init(cn);
+
+            int[] activeColumns = { 4, 5 };
+            Cell[] burstingCells = cn.GetCells(new int[] { 0, 1, 2, 3, });
+
+            ComputeCycle cc = tm.Compute(activeColumns, true) as ComputeCycle;
+
+            Assert.IsFalse(cc.ActiveCells.SequenceEqual(burstingCells));
+        }
+
+        /// <summary>
+        ///Test a  if no cells have active segments, activate all the cells which cant be predicted in columns
+        /// </summary>
+        [TestMethod]
+        [TestCategory("Prod")]
+        public void TestBurstNotpredictedColumns()
+        {
+            TemporalMemory tm = new TemporalMemory();
+            Connections cn = new Connections();
+            Parameters p = getDefaultParameters();
+            p.apply(cn);
+            tm.Init(cn);
+
+            int[] activeColumns = { 1, 2 }; //Cureently Active column
+            IList<Cell> burstingCells = cn.GetCells(new int[] { 0, 1, 2, 3 }); //Number of Cell Indexs
+
+            ComputeCycle cc = tm.Compute(activeColumns, true) as ComputeCycle; //COmpute class object 
+
+            Assert.IsFalse(cc.ActiveCells.SequenceEqual(burstingCells));
+        }
+
+        /// <summary>
+        ///Test a Un-change for non selected matching in no cells have active segments, activate 4 and 5 cell in the column
+        /// </summary>
+
+        [TestMethod]
+        [TestCategory("Prod")]
+        public void TestNoChangeToNoTSelectedMatchingSegmentsInBurstingColumn()
+        {
+            TemporalMemory tm = new TemporalMemory(); // TM class object
+            Connections cn = new Connections();
+            Parameters p = getDefaultParameters(null, KEY.PERMANENCE_DECREMENT, 0.08); // Used Permanence decrement parameter 
+
+            p.apply(cn);
+            tm.Init(cn);
+
+            int[] previousActiveColumns = { 0 };
+            int[] activeColumns = { 1 };
+            Cell[] previousActiveCells = { cn.GetCell(0), cn.GetCell(1), cn.GetCell(2), cn.GetCell(3) };
+            // no cells have active segments, activate 4 and 5 cell in the column
+            Cell[] burstingCells = { cn.GetCell(4), cn.GetCell(5) };
+
+            DistalDendrite selectedMatchingSegment = cn.CreateDistalSegment(burstingCells[0]);
+            cn.CreateSynapse(selectedMatchingSegment, previousActiveCells[0], 0.3);
+            cn.CreateSynapse(selectedMatchingSegment, previousActiveCells[1], 0.3);
+            cn.CreateSynapse(selectedMatchingSegment, previousActiveCells[2], 0.3);
+            cn.CreateSynapse(selectedMatchingSegment, cn.GetCell(81), 0.3);
+
+            DistalDendrite otherMatchingSegment = cn.CreateDistalSegment(burstingCells[1]);
+            Synapse as1 = cn.CreateSynapse(otherMatchingSegment, previousActiveCells[0], 0.3);
+            Synapse is1 = cn.CreateSynapse(otherMatchingSegment, cn.GetCell(81), 0.3);
+
+            tm.Compute(previousActiveColumns, true);
+            tm.Compute(activeColumns, true);
+
+            Assert.AreEqual(0.3, as1.Permanence, 0.01);
+            Assert.AreEqual(0.3, is1.Permanence, 0.01);
+        }
+
+
+        /// <summary>
+        ///Test a active column Where most used cell in a column and after every test its alter the cell
+        /// </summary>
+        [TestMethod]
+        public void TestRandomMostUsedCell()
+        {
+            TemporalMemory tm = new TemporalMemory();
+            Connections cn = new Connections();
+            Parameters p = getDefaultParameters(null, KEY.COLUMN_DIMENSIONS, new int[] { 3 });
+            p = getDefaultParameters(p, KEY.CELLS_PER_COLUMN, 2);
+            p.apply(cn);
+            tm.Init(cn);
+
+            DistalDendrite dd = cn.CreateDistalSegment(cn.GetCell(1));
+            cn.CreateSynapse(dd, cn.GetCell(0), 0.30);
+
+            for (int i = 0; i < 10; i++)
+            {
+                Assert.AreEqual(4, TemporalMemory.GetLeastUsedCell(cn, cn.GetColumn(2).Cells, cn.HtmConfig.Random).Index);
+            }
+        }
+
+        /// <summary>
+        /// Activates all of the cells in an unpredicted active column
+        /// </summary>
+        [TestMethod]
+        [TestCategory("Prod")]
+        public void TestActivatedunpredictedActiveColumn()
+        {
+            HtmConfig htmConfig = GetDefaultTMParameters();
+            Connections cn = new Connections(htmConfig);
+            TemporalMemoryMock tm = new TemporalMemoryMock();
+            tm.Init(cn); ///use connection for specified object to build to implement algoarithm 
+            Random random = cn.HtmConfig.Random;
+            int[] prevActiveColumns = { 1, 2, 3, 4 }; /// 
+            Column column = cn.GetColumn(6); /// Retrieve column 6 
+            IList<Cell> preActiveCells = cn.GetCells(new int[] { 0, 1, 2, 3 }); /// 4 pre-active cells
+            IList<Cell> preWinnerCells = cn.GetCells(new int[] { 0, 1 }); ///Pre- winners cells from pre avtive once
+            List<DistalDendrite> matchingsegments = new List<DistalDendrite>(cn.GetCell(3).DistalDendrites); ///Matching segment from Distal dentrite list
+            var BustingResult = tm.BurstColumnMock(cn, column, matchingsegments,
+                                 preActiveCells, preWinnerCells, 0.10, 0.10,
+                                                new ThreadSafeRandom(100), true);
+            // Assert.AreEqual(, BustingResult);
+            Assert.AreEqual(6, BustingResult.BestCell.ParentColumnIndex);
+            Assert.AreEqual(1, BustingResult.BestCell.DistalDendrites.Count());
+        }
+    }
+
+    public class TemporalMemoryMock:  TemporalMemory
+    {
+        public BurstingResult BurstColumnMock(Connections conn, Column column, List<DistalDendrite> matchingSegments,
+           ICollection<Cell> prevActiveCells, ICollection<Cell> prevWinnerCells, double permanenceIncrement, double permanenceDecrement,
+               Random random, bool learn)
+        {
+            return base.BurstColumn(conn,column,matchingSegments, prevActiveCells, prevWinnerCells, permanenceIncrement, permanenceIncrement, random , learn);
         }
     }
 }
